# Speed up checked kernels for non-null data (~1.4-5x faster)