# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #2747

# Rationale for this change
We can use some more optimized ways to generate new arrays from fallible iterators:
(There is an overflow in the benchmark).
 
```
add_checked(0)          time:   [14.555 µs 14.574 µs 14.596 µs]
                        change: [-79.379% -79.172% -78.968%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

Benchmarking subtract_checked(0): Collecting 100 samples in estimated 5.0209 s (subtract_checked(0)     time:   [15.246 µs 15.265 µs 15.285 µs]
                        change: [-82.424% -81.720% -81.002%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  2 (2.00%) high severe

Benchmarking multiply_checked(0): Collecting 100 samples in estimated 5.0334 s (multiply_checked(0)     time:   [14.364 µs 14.381 µs 14.399 µs]
                        change: [-80.759% -80.340% -79.944%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe

Benchmarking divide_checked(0): Collecting 100 samples in estimated 5.2102 s (96divide_checked(0)       time:   [54.127 µs 54.173 µs 54.223 µs]
                        change: [-29.294% -29.093% -28.909%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe
```
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->