@Dandandan 2022-09-16
Ah wait - this changes the semantics slightly as the overflow check probably isn't meant to work on null-values...
Requires some more thought.